% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bioinfo_fns.R
\name{add_bed_id}
\alias{add_bed_id}
\title{Add a fourth "ID" column to bed-style files}
\usage{
add_bed_id(bed_file)
}
\arguments{
\item{bed_file}{either a dataframe or filename to be read}
}
\description{
There are three required BED fields: chr, chromStart
and chromEnd. The fourth (optional) column is reserved
for name which this function generates as "chr-chromStart".
}
\examples{
bed <- data.frame(chr=c("chr1", "chr2"),
  start=c(100, 450), end=c(200, 300))

add_bed_id(bed)
}

