% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bioinfo_fns.R
\name{bioconductor}
\alias{bioconductor}
\title{Alias to source remote bioconductor script}
\usage{
bioconductor()
}
\description{
Remembering/typing this URL is a pain.
}
\examples{
bioconductor()
biocLite()
}

