% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmr-package.R
\docType{package}
\name{blmr-package}
\alias{blmr}
\alias{blmr-package}
\title{blmr: Bayesian linear model in R.}
\description{
Bayesian linear model in R.
}
\author{
\strong{Maintainer}: Shohei Ogawa \email{s_ogawa@akarf.com}

}
\keyword{internal}
