% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{callStates}
\alias{callStates}
\title{Call compartment state sequence}
\usage{
callStates(indat)
}
\arguments{
\item{any}{Numeric matrix in wide format}
}
\value{
data.frame with eigen and state
}
\description{
Uses an HMM to call the most likely state sequence
from a numeric vector of observed compartment eigenvectors.
}
\examples{
barsWErrors(matrix(rnorm(100, mean=2, sd=.5), ncol=10))
}

