% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{circPermute}
\alias{circPermute}
\title{Circular permutation of a vector}
\usage{
circPermute(vec, n)
}
\arguments{
\item{vec}{vector to permute}

\item{n}{number of positions to shift}
}
\description{
Takes a vector and shifts it \code{n} places,
wrapping shifted cells to effect a "circular permutation".
}
\examples{
circPermute(1:10, 3)
}
\keyword{permute}

