% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/general_fns.R
\name{darken_col}
\alias{darken_col}
\title{Darken an R colour by a specified proportion}
\usage{
darken_col(colname, by = 0.05)
}
\arguments{
\item{by}{proportion to darken colour (default \code{.05} for 5 percent)}
}
\description{
Supply a colour name or RGB and a proportion
between 0 and 1.
}
\examples{
cols <- sapply(seq(0, .5, by=.1), function(i) darken_col("red", i))
image(as.matrix(1:5), col=cols)
}

