% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/3dgenome_specific.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Add linear regression eqn to ggplot2 graph}
\usage{
lm_eqn(m)
}
\description{
Use with e.g. \code{annotate} to add a linear regression
equation and \eqn{R^2}{R^2} value to a plot.
}
\examples{
library("ggplot2")

set.seed(42)
df <- data.frame(x=1:10, y=1:10*rnorm(10))

ggplot(df, aes(x=x, y=y)) + geom_point() +
 geom_smooth(method="lm", se=F) +
 annotate("text", x=4, y=10, label=lm_eqn(lm(y ~ x, df)), parse=T)
}
\references{
SO answer by user Jayden and others: \url{http://stackoverflow.com/a/13451587/1274516}
}

