% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/general_fns.R
\name{match_order}
\alias{match_order}
\title{Order one vector to match another}
\usage{
match_order(...)
}
\arguments{
\item{x}{see `?match`}
}
\description{
There are three required BED fields: chr, chromStart
and chromEnd. The fourth (optional) column is reserved
for name which this function generates as "chr-chromStart".
}
\examples{
a <- 1:5
b <- 5:1

(match_order(a, b))
# [1] 5 4 3 2 1
}

