% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{modelEigens}
\alias{modelEigens}
\title{randomForest modelling of eigenvectors}
\usage{
modelEigens(all.dat, n = 10, ...)
}
\arguments{
\item{dat}{\code{data.frame} in the format \eqn{[Y, X_1, X_2, ... X_n]}{[Y, X_1, X_2, ... X_n]}}

\item{k}{number of folds}
}
\value{
\code{list} of "preds": RF predictions across all folds
  "some other stff"
}
\description{
Later switched to \code{modelEigens.all}. This version creates manual
folds for \eqn{k}{k}-fold cross-validation (default k = 10), manual
as opposed to using \code{caret} methods correctly. Cross-validation
isn't really necessary with a \code{randomForest} approach as each
tree is tested on out-of-bag (OOB) data and this can be averaged over
the forest to reproduce essentially a bootstrapped, many sample
cross-validation scheme.
}
\examples{
modelEigens(h.dat)
}
\seealso{
\code{\link{modelEigens.all}}
}

