% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{panel.pts2}
\alias{panel.pts2}
\title{Panel drop-in for corrgram}
\usage{
panel.pts2(x, y, corr = NULL, col.regions, ...)
}
\description{
Helps with overplotting for high density scatterplots.
Use in e.g. upper.panel=panel.pts2
}
\examples{
library("corrgram")

corrgram(matrix(rnorm(10000), 2000),
  upper.panel=panel.pts2,
  lower.panel=panel.pts)
}

