% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_fns.R
\name{permute_chr}
\alias{permute_chr}
\title{Circular permutation of genomic features in a BED file}
\usage{
permute_chr(chr, bed, chr.sizes, step.size = 1)
}
\arguments{
\item{chr}{chromosome reference, matching your BED file (i.e.
this may be "chr1" or "1")}

\item{bed}{a BED format \code{data.frame}, such as that
produced after using \code{read.table} on a BED file}

\item{chr.sizes}{a \code{data.frame} of chromsoome sizes,
i.e. \code{mm9.chrom.sizes.txt}}

\item{step.size}{if your features are discretised (e.g. can only
occur at set intervals, like those introduced by binning) they
can be permuted by setting \code{step.size} to the size of your
bins in basepairs. \code{step.size} of 1 (the default) means no
binning.}
}
\description{
Generate a permuted version of an input BED file (object),
probably for use as a null model for comparison with some
statistical result.
}
\examples{
sizes <- read.table("http://igv.googlecode.com/svn-history/r2493/trunk/src/org/broad/igv/hic/tools/mm9.chrom.sizes")

bed_file <- data.frame(chr=c("chr1", "chr2"),
  start=c(100, 450), end=c(200, 300))

set.seed(42)

(permute_chr("chr1", bed_file, sizes))
#    chr     start       end
# 1 chr1 180395673 180395773

# run for all chromosomes
chrs <- as.character(unique(bed$chr))
perm <- sapply(chrs, permute_chr, bed=bed_file,
               chr.sizes=sizes, simplify=F)
(do.call(rbind, perm))
#        chr     start       end
#  chr1 chr1 184787091 184787191
#  chr2 chr2  52005764  52005614
}

