% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{plotPredRes}
\alias{plotPredRes}
\title{Scattergraph of model predictions}
\usage{
plotPredRes(modelOut = NA, x = NA, y = NA, col = "blue", ct = "H1",
  scale.factor = 0.7)
}
\arguments{
\item{x}{Predictions (numeric vector)}

\item{y}{Observations (numeric vector)}
}
\description{
Plots empirical vs. obs values with some summary metrics.
Tries to help reduce overplotting using transparency and
overlaid density isoclines.
}
\examples{
plotPredRes(h.mod$predicted, h.mod$y)
}

