% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.blm.R
\name{predict.blm}
\alias{predict.blm}
\title{Predict value by bayesian linear model}
\usage{
\method{predict}{blm}(model, new_data, type = "response")
}
\arguments{
\item{model}{a bayesian model for which prediction is desired.}

\item{new_data}{new data for which prediction is desired.}

\item{type}{the type of prediction required.}
}
\description{
Predict value by bayesian linear model
}
