% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/3dgenome_specific.R
\name{summarySE}
\alias{summarySE}
\title{Build summary data.frame}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\description{
Essentially just wraps a fairly simple \code{plyr} munging call.
Better to just learn \code{dplyr} properly.
}
\references{
\url{http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)/}
}

