% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_blm}
\alias{theme_blm}
\title{blm custom ggplot2 theme}
\usage{
theme_blm(base_size = 12, base_family = "", use_sizes = TRUE)
}
\description{
Custom ggplot2 theme with \url{blm.io} colourscheme.
Reasonably minimalist but retains grid lines which might
be better off stripped depending on data.
}
\examples{
library("ggplot2")
library("grid")

# default theme
ggplot(data.frame(x=1:100, y=rnorm(100)), aes(x, y)) +
  geom_point()

#blm
ggplot(data.frame(x=1:100, y=rnorm(100)), aes(x, y)) +
  geom_point() + theme_blm()
}

