% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/general_fns.R
\name{title_case}
\alias{title_case}
\title{Converts a string to Title Case}
\usage{
title_case(x)
}
\arguments{
\item{x}{string}
}
\description{
Converts a string of any case to Title Case; that is, capitalise
the first letter of each word and lower case the rest. Currently
does not skip words like or, of etc. Edited from an R manual.
}
\examples{
title_case("new York, nEW YorK!")
# "New York, New York!"
}

