% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{barsWErrors}
\alias{barsWErrors}
\title{Barplot with standard errors}
\usage{
barsWErrors(any, ...)
}
\arguments{
\item{any}{Numeric matrix in wide format}
}
\description{
Takes a wide-format matrix and builds a bar chart with standard errors
i.e. used for plotting model results and variable important with k-fold
cross-validation results
}
\examples{
barsWErrors(matrix(rnorm(100, mean=2, sd=.5), ncol=10))
}

