% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{load.all.pkgs}
\alias{load.all.pkgs}
\title{Load all packages for project}
\usage{
load.all.pkgs()
}
\description{
Load all packages used in higher order project. Probably a
good way to test if you have all installed. Aside, could take
a parameter to specify which project you're loading packages for
and what to do if they're missing (try to install?). N.B.
Obviously not all of these packages required in each script.
}
\examples{
\dontrun{blmR::load.all.packages() }
}

