% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/general_fns.R
\name{match_order}
\alias{match_order}
\title{Order one vector to match another}
\usage{
match_order(...)
}
\arguments{
\item{x}{see `?match`}
}
\description{
Simply calls `order(match(...)`.
}
\examples{
a <- letters[5:1]
b <- letters[1:10]

(a[match_order(a, b)])
# [1] "a" "b" "c" "d" "e"

(b[match_order(b, a)])
#  [1] "e" "d" "c" "b" "a" "f" "g" "h" "i" "j"
}

